==> R/01_load.qmd <==
---
title: "Importing Raw data"
format: 
  html:
    self-contained: true
editor: visual
---

# Load Libraries

```{r}
#| message: false
library(here)
```

# Download raw data

```{r}
folder <- here("data", "_raw")
local_path <- file.path(folder, "raw_data.tar.gz")

```

```{r}
if (!dir.exists(folder)) {
  dir.create(folder, recursive = TRUE)
}
download.file("https://cbioportal-datahub.s3.amazonaws.com/tmb_mskcc_2018.tar.gz", local_path, mode = "wb")
```

# Unpack raw data

```{r}
untar(
  tarfile = local_path,
  exdir = folder
)
```
