==> R/04_1_data_clinical_patient_analysis.qmd <==
---
title: "data_clinical_patient_analysis"
format: 
  html:
    self-contained: true
editor: visual
---

# Importing libraries

```{r}
#| echo: false
library("tidyverse")
library('here')
library('survival')
library('survminer')
library('viridis')
```

# Importing data

```{r}
#| message: false
patients_sample_gene_panel <- read_tsv(here("data", "patient_sample_gene_panel.tsv"))
```

# Subsetting Data

From the merge data set, we will do analysis on the patients information for now, so taking only the columns that were originally in the patient data set, diregarding the cancer types,..

```{r}
patients <- patients_sample_gene_panel |> 
  select(c(SEX, OS_MONTHS, OS_STATUS, AGE_GROUP, DRUG_TYPE)) |> 
  mutate(AGE_GROUP = factor(
      AGE_GROUP,
      levels = c("<30", "31-50", "50-60", "61-70", ">71"),
      ordered = TRUE
      )
      )
```

# Plots

First, we will be looking at the age distribution of Cancers in men and woman and separating the patients that died from the living ones.

```{r}
patients |>
  ggplot(aes(x = AGE_GROUP,
             fill = factor(SEX))) +
  geom_bar()+
  scale_fill_viridis(option = "viridis",
                     discrete = TRUE,
                     alpha = 0.7)+
  facet_wrap(
  ~OS_STATUS,
  labeller = labeller(OS_STATUS = 
                        c(`0` = "Alive",
                          `1` = "Deceased")
                      ))+
  labs(title="Proportion of Cancer diagnosed in Men and Woman at different ages and their current status", 
       x = "Age group at Diagnosis (years)",
       fill = 'Gender')+
  theme_minimal()+
  theme(plot.title = element_text(size = 9))
```

In both groups, dead or alive, the patients, men and woman that have a cancer diagnosis are usually between 61-70 years old. Cancers seems to be more diagnosed in man than in woman, at all ages.

Looking more into this age group (61-70), lets consider the drug given against cancers. There is also a comparison of the drug given at all ages, for reference. In both cases, male and female are separated (medications do not affect men and women in the same way).

```{r}
patients |> 
  filter(AGE_GROUP == "61-70") |> 
  ggplot(mapping = aes(x = DRUG_TYPE,
                       fill = SEX))+
  geom_bar()+
  labs(title="Drug given to patients diagnosed with a cancer between 61-70years old, differenciated by male and woman", 
       x = "Drug",
       fill = 'Gender')+
  theme(plot.title = element_text(size = 10))+
  scale_fill_viridis(option = "viridis",
                     discrete = TRUE,
                     alpha = 0.7)

#Comparing it to Patients of All Ages
ggplot(patients, 
       mapping = aes(x = DRUG_TYPE,
                       fill = SEX))+
  geom_bar()+
  scale_fill_viridis(option = "viridis",
                     discrete = TRUE,
                     alpha = 0.7)+
  labs(title="Drug given to patients, of all ages, differenciated by gender", 
       x = "Drug",
       fill = 'Gender')+
  theme_minimal()+
  theme(plot.title = element_text(size = 10))
  
```

Looking at the 61-70 years old again, We want to see the time these patients survived for depending on the drug they were given. Again man and Woman are separated.

```{r}
patients |>
  filter(AGE_GROUP == "61-70") |> 
  ggplot(mapping = aes(x = SEX,
                       y = OS_MONTHS, 
                       fill = DRUG_TYPE))+
  geom_boxplot()+
  scale_fill_viridis(option = "viridis",
                     discrete = TRUE,
                     alpha = 0.7)+
  
  labs(title = 'Box Plot of the Drug success rate based on the months the patients lived, separated by gender',
       x = 'Gender',
       y = 'Time survived (in months)',
       fill = 'Medication')+
  theme_minimal()+
  theme(plot.title = element_text(size = 10))
```

Here patients that had lived and those that are dead are mixed, could be see if there is a distinction, between the number of patients dead or alive, based on the same components of the previous plot.

```{r}
patients |>
  filter(AGE_GROUP == "61-70") |> 
  ggplot(mapping = aes(x = DRUG_TYPE,
                       y = OS_MONTHS,
                       fill = factor(OS_STATUS)))+
  geom_boxplot()+
  scale_fill_viridis(label = c('Alive', "Desceased"),
                     option = "viridis",
                     discrete = TRUE,
                     alpha = 0.7)+
  
  facet_wrap(~ SEX)+
  labs(title = 'Box Plot of the Drug success rate based on the proportion of patients status, separated by gender', 
       fill = 'Status', 
       x = 'Drug Type against Cancer', 
       y = 'Number of months survived')+
  theme_minimal()+
  theme(plot.title = element_text(size = 10))
```

In General, PDL-1 is given more than other drugs, CTL4 is rarely given and sometimes Combo.

As for the 61-70 age group, woman given CTLA4, both survived and died the most.

For both men and women PDL-1 has the lowest death proportion. It could be that PDL-1 is generally more given, in cancers that are less aggressive, and combo or CTLA4 for more aggressive cancers that do not react to PDL-1.

## Survival rate

With the patient data, containing Overall Survival months (OS_MONTHS) and Overall Survival Status (OS_STATUS), it is possible to compute a survival probability over time, using the Kaplan Meier method. The Survival package follows this principal and allows us to compute the survival rate of the patients. To do so, Surv() creates the formula object, associating the status of a patients to it's survival time. Then, survfit() uses the formula object to fit it to the Kaplan Meier estimator, calculating the survival probabilities at each observed event time. The result can then be used to plot the Kaplan–Meier survival.

```{r}

#Using the Kaplan Meier based survival function: 
KM_f <- survfit(Surv(OS_MONTHS, OS_STATUS) ~1, data = patients)

#convert to a tibble

tibble_KM_f <- tibble(
  time = KM_f$time,
  survival = KM_f$surv,
  lower = KM_f$lower,
  upper = KM_f$upper)


ggplot(tibble_KM_f, 
       mapping = aes(x = time,
                     y = survival))+
  geom_step()+
  geom_ribbon(aes(ymin = lower, ymax = upper), alpha = 0.3, color = 'blue') +
  coord_cartesian(ylim = c(0,NA))+
  labs(title = 'Overall Survival rate of the patients', 
       x = "Time (in months)",
       y = 'Overall Survival Probability')+
  theme_minimal()
  




```

From the survival analysis, there is a rapid decline in the survival probability in the first months (20-30).

It would be interesting to look at the impact of the different variables between the patients, mainly if the age has an impact on survival or if the gender could also impact the survival probability. From this data set, we also have the drug that was given to each patients, we can see if a drug is more successful by looking at its survival probability.

**By Gender:**

```{r}

#Using the Kaplan Meier based survival function, declining it to the different sobservations that we can make
#survival rate based on the gender of the patients:


KM_f <- survfit(Surv(OS_MONTHS, OS_STATUS) ~SEX, data = patients)

#convert to a tibble

tibble_KM_f <- tibble(
  time = KM_f$time,
  survival = KM_f$surv,
  strata = rep(names(KM_f$strata),KM_f$strata),
  lower = KM_f$lower,
  upper = KM_f$upper)

tibble_KM_f <- tibble_KM_f |> 
  mutate(strata = str_remove(strata, 'SEX ='))

ggplot(tibble_KM_f,
       mapping = aes(x = time, 
                     y = survival, 
                     color = strata))+
  geom_step()+
  scale_color_viridis_d(label = c('Female', 'Male'), 
                      option = "viridis",
                      alpha = 1)+
  coord_cartesian(ylim = c(0, NA))+
  labs(title = 'Survival Curve of the patients depending on their gender', 
       x ='Time (in Months)',
       y = 'Overall Survival Probability',
       color = 'Gender')+
  theme_minimal()
  
```

Overall, the gender has a minimal impact on the survival probability.

**By Age group:**

```{r}
KM_f <- survfit(Surv(OS_MONTHS, OS_STATUS) ~AGE_GROUP, data = patients)

#convert to a tibble

tibble_KM_f <- tibble(
  time = KM_f$time,
  survival = KM_f$surv,
  strata = rep(names(KM_f$strata),KM_f$strata))

tibble_KM_f <- tibble_KM_f %>%
  mutate(strata = str_remove(strata, "AGE_GROUP="),
         strata = factor(strata,
                         levels = c("<30", "31-50", "50-60", "61-70", ">71"),
                         ordered = TRUE
                         )
         )

ggplot(tibble_KM_f,
       mapping = aes(x = time, 
                     y = survival, 
                     color = strata))+
  geom_step()+
  scale_color_viridis_d(option = "viridis",
                        alpha = 1)+
  coord_cartesian(ylim = c(0, NA))+
  labs(title = 'Survival Curve of the patients depending on their Age at the dignostic', 
       x = 'Time (in Months)',
       y = 'Overall Survival Probability',
       color = 'Age Group')+
    theme_minimal()
```

**By Drug type**

```{r}
KM_f <- survfit(Surv(OS_MONTHS, OS_STATUS) ~DRUG_TYPE, data = patients)

#convert to a tibble

tibble_KM_f <- tibble(
  time = KM_f$time,
  survival = KM_f$surv,
  strata = rep(names(KM_f$strata),KM_f$strata))

tibble_KM_f <- tibble_KM_f %>%
  mutate(strata = str_remove(strata, "DRUG_TYPE="))

ggplot(tibble_KM_f,
       mapping = aes(x = time, 
                     y = survival, 
                     color = strata))+
  geom_step()+
  scale_color_viridis_d(label = c('CTLA4', 'PD-1/PDL-1', 'Combo'))+
  coord_cartesian(ylim = c(0, NA))+
  labs(title = 'Survival Curve of the patients depending on their Medication', 
       x ='Time (in Months)', 
       y = 'Overall Survival Probability',
       color = 'Drug')+
  theme_minimal()
```

Overall, the survival probability increases after 40 months as the slope of the curve is less steep. It is possible that in some cases only a few patients are left then, and were not reported further.
