==> R/04_4_gene_panel_mutation.qmd <==
---
title: "gene_panel_mutation"
format: 
  html:
    self-contained: true
editor: visual
---

## Load libraries:

```{r}
library(tidyverse)
library(here)
library(dplyr)
library(tidyr)
library(ggplot2)
library(viridis)
```

## Data wrangling:

Load the Gene Panel Matrix dataset and select the *SAMPLE_ID* and *IMPACT_version* columns.

```{r}
#Load the gene panel matrix data
gene_panel_matrix <- read_tsv(here("data",
                                     "patient_sample_gene_panel.tsv"))
#Select the column of interest
gene_panel_matrix <- gene_panel_matrix |> 
  select(SAMPLE_ID,
         IMPACT_version)
gene_panel_matrix
```

Load the Mutations dataset.

```{r}
#Load the mutations data
data_mutation <- read_tsv(here("data",
                               "data_mutation_cleaned.tsv"),
                          show_col_types = FALSE)
```

Some rows in the data_mutation object are duplicates that differ only in the "consequence" column. Given that the consequence is not going to be studied, these duplicate rows will be removed.

```{r}
#Remove the duplicates
data_mutation <- data_mutation |> 
  distinct(across(-consequence),
           .keep_all = TRUE)
```

Rename some columns to keep the same format.

```{r}
#Rename the ID column and the gene column
data_mutation <- data_mutation |> 
  rename(SAMPLE_ID = Tumor_Sample_Barcode,
         GENE = Hugo_Symbol)
```

Select the relevant columns from the data_mutation dataset and merge both datasets by SAMPLE_ID.

```{r}
#Select the columns of interest
data_mutation <- data_mutation |> 
  select(SAMPLE_ID,
         GENE,
         Variant_Classification)
#Merge both objects in a new one
combined_data <- left_join(gene_panel_matrix,
                           data_mutation,
                           by = "SAMPLE_ID")
```

Organize the data by grouping it by **SAMPLE_ID**, ensuring that each row corresponds to a single sample with a tibble listing the mutated genes. Then, calculate the number of mutations per patient.

```{r}
#Group the data by the Sample ID and count the number of mutations per patient
combined_data <- combined_data |> 
  group_by(SAMPLE_ID) |> 
  summarise(
    IMPACT_version = first(IMPACT_version),
    n_mutation = n(),
    mutations = list(select(cur_data(),
                            GENE,
                            Variant_Classification))        
  )

combined_data
```

Calculate the percentage of each mutation for each IMPACT version. Group the less represented mutations into broader categories to improve the final plots clarity.

```{r}
#Create a new object with the percentages of each type of mutation per
#IMPACT version
percentage_type <- combined_data |> 
  select(IMPACT_version,
         mutations) |> 
  #It is necessary to temporary unnest to work with the nested tibble data
  unnest(mutations)

#Group some mutation types to reduce the number of different categories
percentage_type <- percentage_type |>
  #Combine Frame shift del and ins into "Frame_Shift_Mutation"
  mutate(Variant_Classification = case_when(
    Variant_Classification %in% c(
      "Frame_Shift_Del",
      "Frame_Shift_Ins") ~ "Frame_Shift_Mutation",
    TRUE ~ Variant_Classification)) |>
  #Group the less represented mutations into "Other"
  mutate(Variant_Classification = ifelse(
    Variant_Classification %in% c("Frame_Shift_Mutation",
                                  "Missense_Mutation",
                                  "Nonsense_Mutation",
                                  "Splice_Site"),
    Variant_Classification,
    "Other"
  )) |> 
  #Calculate the percentage
  count(IMPACT_version,
        Variant_Classification) |> 
  group_by(IMPACT_version) |> 
  mutate(percentage = n / sum(n) * 100)

#Order the categories for a better representation
percentage_type <- percentage_type |> 
  mutate(Variant_Classification = factor(Variant_Classification,
                                         levels = c("Missense_Mutation",
                                                    "Nonsense_Mutation",
                                                    "Frame_Shift_Mutation",
                                                    "Splice_Site",
                                                    "Other")))
```

## Data representation

### Number of mutations per IMPACT version

A representation of the number of mutations detected by the IMPACT version is done in order to assess whether the latest versions of IMPACT demonstrate improved capacity for identifying new mutations.

```{r}
mutation_count_IMPACT <- ggplot(data = combined_data,
                           mapping = aes(x = IMPACT_version,
                                         y = n_mutation,
                                         fill = IMPACT_version)) +
                      geom_boxplot(alpha = 0.5) +
                      #It is necessary to do the log10 of the y values to 
                      #prevent outliers from hiding the main distribution
                      scale_y_log10() +
                      theme_minimal() +
                      theme(legend.position = "none") +
                      scale_fill_viridis(option = "viridis",
                                         discrete = TRUE) +
                      labs(title = "Number of mutations per IMPACT version (log scale)",
                           y = "Mutations (log10)",
                           x = "IMPACT version")
mutation_count_IMPACT
```

As shown in the plot, the overall number of mutations detected increases with higher IMPACT versions, reflecting that a greater number of studied genes corresponds to a higher number of detected mutations. The medians of detected mutations are the following:

```{r}
results <- combined_data |> 
  group_by(IMPACT_version) |> 
  summarise(mut_median = median(n_mutation))

results
```

### Distribution of mutation type by IMPACT version

```{r}
ggplot(data = percentage_type,
       mapping = aes(x = IMPACT_version,
                     y = percentage,
                     fill = Variant_Classification)) +
  geom_col(alpha = 0.8) + 
  theme_minimal() + 
  theme(panel.grid.major.x = element_blank()) +
  scale_fill_viridis(option = "viridis",
                     discrete = TRUE) +
  labs(title = "Distribution of mutation type by IMPACT version",
       y = "Percentage",
       x = "IMPACT version",
       fill = "Mutation type")
```

While newer IMPACT versions detect more mutations overall, the distribution of mutation types remains largely unchanged. The most significant change is the proportion of Frameshift mutations, which increases with each IMPACT version.

This indicates that version upgrades increase sensitivity without substantially altering the mutation distribution.

## Saving figures:

```{r}
ggsave("../results/figures/mutation_count_IMPACT.png",
       plot = mutation_count_IMPACT)
```
