==> R/06_mutation_types.qmd <==
---
title: "mutation_types"
format: 
  html:
    self-contained: true
editor: visual
---

## Load libraries

```{r}
#| message: False
library(tidyverse)
library(here)
library(dplyr)
library(tidyr)
library(ggplot2)
library(viridis)
```

## Data wrangling

Load the mutations dataset and remove duplicate entries that only differ in the *consequence* column (they are not going to be studied). Select the columns of interest and rename them to *SAMPLE_ID*, *GENE* and *Variant_Classification*. Finally, group the less represented mutation types into broader categories to improve the final plots clarity.

```{r}
#Load the mutations data
mutations <- read_tsv(here("data",
                             "data_mutation_cleaned.tsv"))
#Remove the duplicates
mutations <- mutations |> 
  distinct(across(-consequence),
           .keep_all = TRUE) |> 
  #Select the columns of interest
  select(Hugo_Symbol,
         Tumor_Sample_Barcode,
         Variant_Classification) |> 
  #Modify the columns name
  rename(SAMPLE_ID = Tumor_Sample_Barcode,
         GENE = Hugo_Symbol) |> 
  #Group some mutation types to reduce the number of different categories
  #Combine Frame shift del and ins into "Frame_Shift_Mutation"
  mutate(Variant_Classification = case_when(
    Variant_Classification %in% c(
      "Frame_Shift_Del",
      "Frame_Shift_Ins") ~ "Frame_Shift_Mutation",
    TRUE ~ Variant_Classification)) |>
  #Group the less represented mutations into "Other"
  mutate(Variant_Classification = ifelse(
    Variant_Classification %in% c("Frame_Shift_Mutation",
                                  "Missense_Mutation",
                                  "Nonsense_Mutation",
                                  "Splice_Site"),
    Variant_Classification,
    "Other"
  ))

mutations
```

Load the Clinical Sample data.

```{r}
#Load the clinical sample data
clinical_sample <- read_tsv(here("data",
                                   "patient_sample_gene_panel.tsv"))
clinical_sample
```

Create a subset to study the differences for each Glioma detailed cancer type.

```{r}
#Create a subset for Glioma detailed cancer types
sample_glioma <- clinical_sample |>  
  filter(CANCER_TYPE=='Glioma') |> 
  select(SAMPLE_ID,
         CANCER_TYPE,
         CANCER_TYPE_DETAILED)

sample_glioma
```

Select the columns of interest in the Clinical Sample dataset.

```{r}
clinical_sample <- clinical_sample |> 
  #Select the columns of interest
  select(SAMPLE_ID,
         CANCER_TYPE)
```

Merge the Clinical Sample dataset with the Mutations dataset. From the resulting object, create a subset for the Glioma subcategories.

```{r}
#Merge the objects
clinical_sample_mutation <- left_join(mutations,
                                      clinical_sample,
                                      by = "SAMPLE_ID")
#Create a subset for glioma mutations
glioma_mutations <- clinical_sample_mutation |> 
  filter(CANCER_TYPE == "Glioma")

glioma_sample_mutation <- left_join(glioma_mutations,
                                    sample_glioma,
                                    by = "SAMPLE_ID")
```

Count the frequency of each mutation across cancer types and within each glioma subcategory in the subset.

```{r}
#Count the frequence of each mutation for each cancer type
mutations_count <- clinical_sample_mutation |> 
  group_by(CANCER_TYPE,
           Variant_Classification) |> 
  summarise(count = n(),
            .groups = "drop")
```

```{r}
#Count the frequence of each mutation for each detailed glioma type
glioma_mutations_count <- glioma_sample_mutation |> 
  group_by(CANCER_TYPE_DETAILED,
           Variant_Classification) |> 
  summarise(count = n(),
            .groups = "drop") |> 
  filter(!is.na(CANCER_TYPE_DETAILED))
```

## Data representation

```{r}
mutation_by_cancer <- ggplot(data = mutations_count,
                         mapping = aes(x = CANCER_TYPE,
                                       y = count,
                                       fill = Variant_Classification)) + 
                    geom_col(alpha = 0.8) +
                    theme_minimal() +
                    theme(axis.text.x = element_text(angle = 90,
                                                     hjust = 1,
                                                     size = 7),
                          panel.grid.major.x = element_blank()) +
                    scale_fill_viridis(option = "viridis",
                                       discrete = TRUE) +
                    labs(title = "Distribution of Mutation Type by Cancer Type",
                         x = "Cancer Type",
                         y = "Mutation Count",
                         fill = "Mutation Type")
```

```{r}
ggplot(data = glioma_mutations_count,
       mapping = aes(x = CANCER_TYPE_DETAILED,
                     y = count,
                     fill = Variant_Classification)) + 
  geom_col(alpha = 0.8) +
  theme_minimal() +
  theme(axis.text.x = element_text(angle = 90,
                                   hjust = 1,
                                   size = 7),
        panel.grid.major.x = element_blank()) +
  scale_fill_viridis(option = "viridis",
                     discrete = TRUE) +
  labs(title = "Distribution of Mutation Type by Glioma Subcategory",
       x = "Detailed Glioma Type",
       y = "Mutation Count",
       fill = "Mutation Type")
```

## Saving figures:

```{r}
ggsave("../results/figures/mutation_by_cancer.png",
       plot = mutation_by_cancer)
```
