==> R/09_gene_enrichment_analysis.qmd <==
---
title: "gene_enrichment_analysis"
format: 
  html:
    self-contained: true
---

# Load Libraries

```{r}
#| message: false
library(tidyverse)
library(here)
library(fgsea)
library(msigdbr)
```

# Load data

## Gene Ontology Information

This loads a dataset where genes with their gene symbol and eEnsemblID are mapped to biological processes. With this information we can see if their are certain processes which are more often targeted by mutations than is expected by chance.

```{r}
BP_df = msigdbr(species = "human", category = "C5", subcategory = "BP")
BP_list = split(BP_df$ensembl_gene, BP_df$gs_name)
```

## Cleaned data

```{r}
#| message: false
patient_sample <- read_tsv(file = here("data","patient_sample_gene_panel.tsv"))
sv <- read_tsv(file = here("data","sv_cleaned.tsv"))
mutations <- read_tsv(file = here("data","data_mutation_cleaned.tsv"))
```

# GSEA with mutations per Gene

The Genes are ranked on the number of mutations or structural variances that have been identified in them.

## SV

Gene Set Enrichment analysis is first applied on to the mutations which cause structural variances.

Data format: Genes + number of SVs

```{r}
clinical_clean <- patient_sample |> select(c(SAMPLE_ID, CANCER_TYPE, SAMPLE_TYPE, TMB_NONSYNONYMOUS))

gene_sv_count <- sv |> 
  left_join(clinical_clean, by = c("Sample_Id" = "SAMPLE_ID")) |>
  filter(CANCER_TYPE == "Glioma")|>
  pivot_longer(cols = c("Site1_Hugo_Symbol", "Site2_Hugo_Symbol"), names_to = "Position", values_to = "Gene_Name")|>
  filter(!is.na(Gene_Name))|>
  count(Gene_Name, sort = TRUE)|>
  left_join(BP_df |> select(gene_symbol, ensembl_gene), by = c("Gene_Name" = "gene_symbol"))|>
  distinct()|>
  na.omit()

gene_sv_stats <- gene_sv_count$n
names(gene_sv_stats) = gene_sv_count$ensembl_gene
length(gene_sv_stats)
```

```{r}
pathway_scoring_gene_sv <- fgseaMultilevel(pathways= BP_list, stats = gene_sv_stats,minSize = 0, maxSize = 500, scoreType = "std")
pathway_scoring_gene_sv <- arrange(pathway_scoring_gene_sv, pval)
head(pathway_scoring_gene_sv |> select(pathway, pval, padj), n = 15)
```

Looking at the pvalue and adjusted p-value these results are not very robust. This is probably due to the small sample size of only 28 genes.

## Mutations

The same analysis, but now with the number of mutations

```{r}
gene_mutation_count <- mutations |> 
  left_join(clinical_clean, by = c("Tumor_Sample_Barcode" = "SAMPLE_ID"))|>
  filter(CANCER_TYPE == "Glioma")|> 
  count(Hugo_Symbol, sort = TRUE)|>
  left_join(BP_df |> select(gene_symbol, ensembl_gene), by = c("Hugo_Symbol" = "gene_symbol"))|>
  distinct()|>
  na.omit()

gene_mutation_stats <- gene_mutation_count$n
names(gene_mutation_stats) = gene_mutation_count$ensembl_gene
```

```{r}
pathway_scoring_gene_m <- fgseaMultilevel(pathways= BP_list, stats = gene_mutation_stats,minSize = 15, maxSize = 500, scoreType = "std")
pathway_scoring_gene_m <- arrange(pathway_scoring_gene_m, padj)
head(pathway_scoring_gene_m|> select(pathway, pval, padj), n = 15)
pathway_scoring_gene_m |>
  filter(padj < 0.05) |> 
  nrow()
```

These results are more reliable than the ones of the structural variances. The most significant pathways also seem logical as they show that pathways are targeted by mutations which influence cell death and also cell cycle processes. This is very common in tumors.

## Comparison of results:

```{r}
overlap <- inner_join(
  pathway_scoring_gene_sv |> head(n=100),
  pathway_scoring_gene_m |>  head(n=100), by = "pathway"
  ) |> 
  select(pathway, pval.x, pval.y)

overlap
```

This compares the 100 most mutated pathways form both analysis. Some interesting Pathways are shown here, for example the behaviour pathways, as it is common that brain cancer patients show changes in behaviour.

# FSC with gene mutation per sample

This is the same analysis as before, but here it is counted in how many samples the gene is mutated. This can alter some counts, if a sample contains multiple mutation per gene.

## SV

```{r}
clinical_clean <- patient_sample |> select(c(SAMPLE_ID, CANCER_TYPE, SAMPLE_TYPE, TMB_NONSYNONYMOUS))

sample_sv_count <- sv |> 
   select(c(Sample_Id, Site1_Hugo_Symbol, Site1_Chromosome, Site2_Hugo_Symbol, Site2_Chromosome, Class, SV_Length))|>
  left_join(clinical_clean, by = c("Sample_Id" = "SAMPLE_ID")) |>
  filter(CANCER_TYPE == "Glioma")|>
  pivot_longer(cols = c("Site1_Hugo_Symbol", "Site2_Hugo_Symbol"), names_to = "Position", values_to = "Gene_Name")|>
  filter(!is.na(Gene_Name))|>
  select(Sample_Id, Gene_Name)|>
  distinct()|>
  count(Gene_Name, sort = TRUE)|>
  left_join(BP_df |> select(gene_symbol, ensembl_gene), by = c("Gene_Name" = "gene_symbol"))|>
  distinct()|>
  na.omit()

sample_sv_stats <- sample_sv_count$n
names(sample_sv_stats) = sample_sv_count$ensembl_gene

pathway_scoring_sample_sv <- fgseaMultilevel(pathways= BP_list, stats = sample_sv_stats,minSize = 0, maxSize = 500, scoreType = "std")
pathway_scoring_sample_sv <- arrange(pathway_scoring_sample_sv, pval)
head(pathway_scoring_sample_sv|> select(pathway, pval, padj), n =15)
```

## Mutations

```{r}
mutation_gene_sample <- mutations |> 
  left_join(clinical_clean |> select(SAMPLE_ID, CANCER_TYPE), by = c("Tumor_Sample_Barcode" = "SAMPLE_ID")) |>
  filter(CANCER_TYPE == "Glioma")|>
  select(Tumor_Sample_Barcode, Hugo_Symbol) |> 
  distinct() |> 
  count(Hugo_Symbol, sort = TRUE) |> 
  left_join(BP_df |> select(gene_symbol, ensembl_gene), by = c("Hugo_Symbol" = "gene_symbol"))|>
  distinct()|>
  na.omit()

gene_mutation_stats <- mutation_gene_sample$n
names(gene_mutation_stats) = mutation_gene_sample$ensembl_gene

pathway_scoring_sample_m <- fgseaMultilevel(pathways= BP_list, stats = gene_mutation_stats,minSize = 15, maxSize = 500, scoreType = "std")
pathway_scoring_sample_m <- arrange(pathway_scoring_sample_m, padj)
head(pathway_scoring_sample_m |> select(pathway, pval, padj), n=15)
pathway_scoring_sample_m |>
  filter(padj < 0.05) |> 
  nrow()
```

The comparison of the both counting versions, result in very similar outputs.

# Comparisons of counting Methods

```{r}
overlap <- inner_join(
  pathway_scoring_gene_m |> filter(padj <0.05),
  pathway_scoring_sample_m |>  filter(padj < 0.05), by = "pathway"
  ) |> 
  select(pathway, pval.x, pval.y)

overlap
```
